search: remove query/alert.go
